requirements.txt
Django
djangorestframework
django-cors-headers
djangorestframework_simplejwt
python-dotenv
